ref(ui): applied feedbacks
