+++ src/sentry/static/sentry/app/components/events/interfaces/keyValueList/types.ts
@@ -1,0 +1,8 @@
+import {Meta} from 'app/types';
+
+export type KeyValueListData = {
+  key: string;
+  subject: React.ReactNode;
+  value: string | null;
+  meta: Meta;
+};
